Renaming Layers.OneHanded and Layers.TwoHanded
These are being renamed so that their meaning is more clear; these layers do not represent OneHanded or TwoHanded items properly, and they do not represent Left or Right hands properly. They are arbitrary item slots for the hands. The rendering and client logic previously looked at these values from conflicting perspectives, and as previously named, the meaning conflicted with the data in practice.